Route BARMAN tool-agent commands before code execution
