--- .github/workflows/patch-barman-tool-agent-router-once.yml
@@ -1,55 +1,0 @@
-name: Patch BARMAN Tool Agent Router Once
-
-on:
-  push:
-    branches:
-      - fix/barman-tool-agent-router-20260902
-    paths:
-      - .github/workflows/patch-barman-tool-agent-router-once.yml
-
-permissions:
-  contents: write
-
-jobs:
-  patch:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v7
-        with:
-          ref: fix/barman-tool-agent-router-20260902
-          fetch-depth: 0
-      - uses: actions/setup-node@v7
-        with:
-          node-version-file: package.json
-          package-manager-cache: false
-      - name: Install locked dependencies
-        run: npm ci --no-audit --no-fund
-      - name: Apply router patch
-        run: node scripts/patch-barman-tool-agent-router-once.mjs
-      - name: Diagnose generated router safely
-        shell: bash
-        run: |
-          set -euo pipefail
-          grep -nE 'const raw=|lineBreak|normalized=|const goals' api/barman-tool-agent-broker.js || true
-          node --input-type=module <<'NODE'
-          import { routeToolAgentCommand } from './api/barman-tool-agent-broker.js';
-          const slash=String.fromCharCode(92);
-          const cases=[
-            ['real-newline','1. أصلح BAR-12'+String.fromCharCode(10)+'2. اربط واتساب'],
-            ['one-backslash','1. أصلح BAR-12'+slash+'n2. اربط واتساب'],
-            ['two-backslashes','1. أصلح BAR-12'+slash+slash+'n2. اربط واتساب'],
-          ];
-          for(const [label,value] of cases) console.log(label,JSON.stringify(value),routeToolAgentCommand(value));
-          NODE
-      - name: Verify focused tests
-        run: node --test test/barman-persistent-tool-agent.test.mjs
-      - name: Commit router and remove one-shot helpers
-        run: |
-          set -euo pipefail
-          rm .github/workflows/patch-barman-tool-agent-router-once.yml scripts/patch-barman-tool-agent-router-once.mjs
-          git config user.name 'barmanai'
-          git config user.email 'barmanai@users.noreply.github.com'
-          git add api/barman-tool-agent-broker.js scripts/barman-tool-agent.mjs test/barman-persistent-tool-agent.test.mjs .github/workflows/patch-barman-tool-agent-router-once.yml scripts/patch-barman-tool-agent-router-once.mjs
-          git commit -m 'Route BARMAN tool-agent commands before code execution'
-          git push origin HEAD:fix/barman-tool-agent-router-20260902
-      # rerun marker: preserve raw newlines v6
